* Fixed bug in libhaggle when haggle_handle_free is called from the event-loop thread when running async

+++ android/PhotoShare/src/org/haggle/PhotoShare/InterestView.java
@@ -86,7 +86,9 @@ public class InterestView extends Activity {
     public boolean onContextItemSelected(MenuItem item) {
         AdapterContextMenuInfo info = (AdapterContextMenuInfo) item.getMenuInfo();
      
-        deleteInterest(info.position);
+        if (item.getTitle() == "Delete") {
+        	deleteInterest(info.position);
+        }
         
         return true;
     }
